Add onCluster callback
Add a callback that gets fired when markers are combined in to a
cluster.

The motivation for this addition is to enable hiding of a Marker popup
when the relvant marker gets clustered. The popup is displayed in a
separate flutter_map layer. I tried adding a popup in the relevant
marker's build() using a Stack however that makes the popup unclickable
because it is outside of the bounds of the parent Stack (this is a known
issue/behaviour https://github.com/rorystephenson/flutter_map_marker_cluster/commit/51486399b45d04658383479eeb49df38cca3fc01).